Add ability to create new room
This allows people on the home page to click a button to
create a new room.

This uses a post request to the rooms API, which creates a
new room using the name generator. Name clashes are resolved
by generating names until one doesn't clash. This...
won't work for long, but it'll do for now.

Rooms are stored in memory. This won't scale forever but
it'll do for now.