Add ability to join and leave rooms
This adds the ability to join and leave rooms to the application.

Allows players to open multiple browser windows and receive events
in all of them, whilst still only playing as a single player.

Leaving a room removes sockets. Once all sockets for a player
have been removed, the player is removed.