Add the ability to vote
This adds some client code that can send the vote to the server,
server code that handles the event and applies it to the room,
server code that then broadcasts the change to all clients,
and client code that handles this broadcast to display the
votes.

Votes are currently shown to everyone as they're made, but
what we actually want is for them to be hidden in the UI until
we're ready to reveal votes. This could be handled in the client,
but it's probably safer to not send the votes to the client until
they're ready to be shown.

We also need a way to see who has voted.