Add self identification
Adds a dinosaur icon displayed next to the player name
to indicate which player you are.

Requires removing the HttpOnly flag from the cookie to allow
it to be inspected from Javascript. This is not really "secure"
but this isn't a secure session, so I don't think it's a big
deal. We'd do something different if we required login
but for now this should do.

Also sets the max age of the cookie to infinity, which will mean
players don't rotate their names every week like they do currently.