Update room to accept Player instance
We "need" a player for some methods but not others, so what
ended up happening was the PokerApp constructed a player, and
then passed in just the ID where only the ID is required.

This works fine, but I think it makes a bit more sense and is a
bit more flexible if we pass in the entire player object from
the consumer side, rather than have the consumer know when
to pass a player and when to pass the ID (or more precisely, a
raw string). This gives us a bit of extra type safety too.