Fix NaN error
When calculating the mean of no votes, we're dividing by zero which
results in a NaN result.

This is correct, but NaN isn't allowed in JSON. This breaks the
serialisation and crashes the connection.

We probably shouldn't care about that in the statistics class
itself as NaN is technically true, but I don't think the distinction
is important. Probably.