Add CDK resources
This adds all the resources required for the poker app using the CDK.

It's pretty much just the EC2 instance, and then anything that it needs to be published.

We're using the "L2" resources in the CDK, which give sensible defaults and
aren't a 1 to 1 mapping with a CFT. This means that we might be
creating extra resources we wouldn't normally (like IAM roles) here as it's
easier to do so now than it was before.