Add cat mode
This adds the ability for a player to turn on "cat mode" by typing "catmode"
on their keyboard when on a page.

Cat mode changes a bunch of the dinosaur themed elements to be cat themed
instead, like the logos, favicons, and card images.

This requires changes on the server, so we can change the "me"
icon to a cat when catmode is on.

This means that catmode is stored on the server, but we don't do much other
than the player icon with this information, and rely on changing the elements
when cat mode is turned on. This turns itself off when refreshing the page
or navigating.

We could use the state of catmode in the session cookie to preserve
catmode across different pages, but that would require more work and
probably isn't required, but we'll see what happens.

I've taken the liberty of compressing the images at the same time,
so hopefully I remember to do that in the future too.