Fix catmode path replacement
The `dinosaur` replacement was replacing the dinosaur in the path,
meaning that it would try to get the images from developmentandcats.co.uk,
which doesn't exist (yet).

This now replaces hyphens and slashes too, so it won't match the
domain name. Could probably fix it a different way, but this will
work fine, probably.