Add deletion delay
This adds a 10-minute delay before a room is deleted once it is
marked for deletion.

This will allow people to quickly "nip back in" to a room without
having to create another.

I was considering having rooms auto-create on a GET request, but
something about idempotency made me reconsider.