Deploy a versioned jar
The CDK isn't picking up any changes as the CFT we produce is the
same as the last time we deployed.

By including the jar version in the CDK config, we'll have a different
CFT and therefore actually deploy. In theory.

I doubt this is the smartest way to do this, but it probably works.