Add version string to JS and CSS files
This adds the deployed version of the service as a query parameter
to the CSS and JS files.

This will help us invalidate the cache when pushing a change, as
the files will have a new name and thus pull a new file, where before
they would pull the cached version.

We could be smarter here and tag the CSS and JS with a checksum on the
file instead, or a last modified date, or something like that.

I don't think we need to do that though - and this also gives me a neat
way to make sure my deployment has actually worked as expected. Although
maybe I should put that on a visible page somewhere? Hmm. This will
do for now.